Moved SetTestOptions to AdaptiveDialogBot.SetUpTurnStateAsync
